return error if repair block failed
increase timeout in TestBroadcastBlock

increase open file limits

fix overflow in write cache parent tracking

+++ lib/trie/database.go
@@ -106,7 +106,7 @@ type cachedNode struct {
 	node node
 	size uint16
 
-	parents  uint16
+	parents  uint32
 	children map[common.Hash]uint16
 
 	flushPrev common.Hash
